projects/dft_lammps_research/platform/api/requirements.txt
# API Platform Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate Limiting
slowapi>=0.1.9
redis>=5.0.0

# HTTP Client (for webhooks and external APIs)
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Utilities
structlog>=23.2.0
python-dotenv>=1.0.0
orjson>=3.9.0

# SDK Dependencies
# These are included for the example scripts
